fixed zip skip folder, files

+++ 3S_AUTO/UploadFileZip_Bin_Grade.py
@@ -29,12 +29,9 @@ def Achive_Folder_To_ZIP(sFilePath, dest = "", sSequenceNumber = "0"):
     sYear  = time.strftime("%Y")
     sDate  = time.strftime("%d")
 
-    
-    sRemoteFileName = '{}{}.{}{}{}{}{}'.format('v1.0.', sYear, iMonth, sDate ,'_Temp', sSequenceNumber,'.zip')
-    sRemoteFolderName = '{}{}.{}{}{}{}'.format('v1.0.', sYear, iMonth, sDate ,'_Temp', sSequenceNumber)
 
-    #sDate = time.strftime("%Y.%m%d")
-    #sRemoteFileName = '{}{}{}{}{}'.format('v1.0.', sDate,'_Temp', sSequenceNumber,'.zip')
+    sRemoteFileName = '{}.{}.{}.{}{}{}'.format('1', sSequenceNumber, sYear, iMonth, sDate ,'.zip')
+    sRemoteFolderName = '{}.{}.{}.{}{}'.format('1', sSequenceNumber, sYear, iMonth, sDate )
 
     dest = os.path.join(dest, sRemoteFileName)         
     #dest = os.path.join(dest, sRemoteFolderName)         
@@ -58,21 +55,28 @@ def Achive_Folder_To_ZIP(sFilePath, dest = "", sSequenceNumber = "0"):
    # sZipFolder = 
     #print sFilePath
     for root, folders, files in os.walk(".\\"):
+        if ( 'Release' in folders ):
+            folders.remove('Release')                 
+            print("Achive_Folder_To_ZIP skip Release", folders)  
+        if ( '.git' in folders ):
+            folders.remove('.git') 
+            print("Achive_Folder_To_ZIP skip .git", folders)         
+        if ( 'workingTMP' in folders ):
+            folders.remove('workingTMP')                 
+            print("Achive_Folder_To_ZIP skip workingTMP", folders)          
+
         for sfile in files:
             stmp = os.path.join(root, sfile) 
 
-            if ( '.git' in folders ):
-                folders.remove('.git') 
-                print("Achive_Folder_To_ZIP remove ", stmp)
-            if ( 'workingTMP' in folders ):
-                folders.remove('workingTMP')                 
-                print("Achive_Folder_To_ZIP remove ", stmp)
             if ('.gitignore' in files):              
                 files.remove('.gitignore')                 
-                print("Achive_Folder_To_ZIP remove ", stmp)            
-            if ('reference.txt' in files):              
+                print("Achive_Folder_To_ZIP skip .gitignore", stmp)            
+            elif ('reference.txt' in files):              
                 files.remove('reference.txt')                 
-                print("Achive_Folder_To_ZIP remove ", stmp)                                                    
+                print("Achive_Folder_To_ZIP skip ", stmp)                                                    
+            elif ('ChangeList.txt' in files):              
+                files.remove('ChangeList.txt')                 
+                print("Achive_Folder_To_ZIP skikp ChangeList.txt", stmp)                                                                    
             else:
                 aFile = os.path.join(root, sfile)
                 sTmpPath = os.path.join(sRemoteFolderName, aFile)
@@ -91,10 +95,10 @@ def removeFolder(sPath):
         print("remove folder = ", sDestinationPath)
         shutil.rmtree(sDestinationPath)
 
-    sDestinationPath = os.path.join(sPath,"src\Release")
-    if (os.path.isdir(sDestinationPath)):
-        print("remove folder = ", sDestinationPath)
-        shutil.rmtree(sDestinationPath)    
+    # sDestinationPath = os.path.join(sPath,"src\Release")
+    # if (os.path.isdir(sDestinationPath)):
+    #     print("remove folder = ", sDestinationPath)
+    #     shutil.rmtree(sDestinationPath)    
 
     sDestinationPath = os.path.join(sPath,"src\Setting")
     if (os.path.isdir(sDestinationPath)):
@@ -138,25 +142,33 @@ def removeFile(sPath):
             os.remove(currentFile)
 
 def copyTo3800(src_file, det_file, sSequenceNumber):
-
-
     datetime.datetime.now()
     sMonth = time.strftime("%m") 
     iMonth = int(sMonth)
     sYear  = time.strftime("%Y")
     sDate  = time.strftime("%d")
-    
-    sRemoteFileName = '{}{}.{}{}{}{}{}'.format('v1.0.', sYear, iMonth, sDate ,'_Temp', sSequenceNumber,'.zip')
-    sRemoteFolderName = '{}{}.{}{}{}{}'.format('v1.0.', sYear, iMonth, sDate ,'_Temp', sSequenceNumber)
 
-    #sDate = time.strftime("%Y.%m%d")
-    #sRemoteFolderName = '{}{}{}{}'.format('v1.0.', sDate,'_Temp', sSequenceNumber)
-    #print("copyTo3800 path = ", sRemoteFolderName)
-    
+
+    sRemoteFolderName = '{}.{}.{}.{}{}'.format('BIN_GRADE_v1', sSequenceNumber, sYear, iMonth, sDate)
+    #sRemoteFolderName = '{}.{}.{}.{}{}'.format('1', sSequenceNumber, sYear, iMonth, sDate )
+
     det_file = os.path.join(det_file, sRemoteFolderName) 
     #print(det_file)
+
+
     if os.path.exists(det_file):
-        shutil.rmtree(det_file)
+        print('{}{}'.format(det_file, "   exist !! remove it ?"))
+        sYesNo = rawInputTest()
+        if (sYesNo == 1):
+            shutil.rmtree(det_file)
+            #os.remove(dest)
+            print('{}{}'.format(det_file, ", rmtree ok"))
+        elif(sYesNo == 0):
+            print("skip")
+            sys.exit(0)
+    else:        
+        print('{}{}'.format(det_file, "  do not exist"))
+        #shutil.rmtree(det_file)
         
     shutil.copytree(src_file, det_file)
 
@@ -186,26 +198,48 @@ def rawInputTest():
 if __name__ == "__main__":
 
     sVersion = 1
-    sProgramimgPath ="D:\\3S_PC\sourceCode\SSD\MP_UI\source_code\GIT_MP_UI_BIN_GRADE\BIN_GRADE_V1.0"
-    sFileServer = r"\\fileserver\Dep_AP\Project\SSD\temp\BinGrade"
-    sRemotePath = r"\\fileserver\3800\SW\SSD_BIN_GRADE"
+    sProgramimgPath = "D:\\3S_PC\sourceCode\SSD\MP_UI\VC6\BIN_GRADE_V1.0"
+    sServerSourceCodePath = r"\\fileserver\Dep_AP\Project\SSD\tools"
+    sServerBinaryPath = r"\\fileserver\3800\SW\tools"
 
-    # sReleaseFile = os.path.join(sProgramimgPath,"src\\Release")
-    # sBinFolder = os.path.join(sProgramimgPath,"bin")
-    # shutil.copy2(sReleaseFile,sBinFolder)
+    #prepare enviroment (sourcecode)
+    sServerSourceCode = os.path.join(sServerSourceCodePath,"BIN_GRADE")
+    if not (os.path.isdir(sServerSourceCode)):
+        os.mkdir(sServerSourceCode)
+        print('{}{}'.format("mkdir done!  ", sServerSourceCode))
 
-    sCleanPath = os.path.join(sProgramimgPath,"src")
-    removeFolder(sCleanPath)
+    sServerSourceCode = os.path.join(sServerSourceCode,"SOURCE_CODE")
+    if not (os.path.isdir(sServerSourceCode)):
+        os.mkdir(sServerSourceCode)
+        print('{}{}'.format("mkdir done!  ", sServerSourceCode))        
 
-    #sCleanPath = os.path.join(sProgramimgPath, "src") 
-    removeFile(sCleanPath)
+    #prepare enviroment (binary file)
+    sServerBinary = os.path.join(sServerBinaryPath,"BIN_GRADE")
+    if not (os.path.isdir(sServerBinary)):
+        os.mkdir(sServerBinary)
+        print('{}{}'.format("mkdir done!  ", sServerBinary))
 
-    
+
+    sSrcReleaseFile = os.path.join(sProgramimgPath,"src\Release\BIN_GRADE.exe")
+    sDstBinFile = os.path.join(sProgramimgPath,"src\\bin")
+    copyReleaseNote(sSrcReleaseFile, sDstBinFile)
+
+
     print("execute UAC !!!")
     sUACPath = os.path.join(sProgramimgPath,"src\\UAC")
     os.chdir(sUACPath)
     os.system("uac_path.bat")
     print('{}{}{}'.format("check done!  ", sUACPath, " SSDMP.exe modify date"))
+
+    
+
+
+    sCleanPath = sProgramimgPath #os.path.join(sProgramimgPath,"src")
+    removeFolder(sCleanPath)
+
+    #sCleanPath = os.path.join(sProgramimgPath, "src") 
+    removeFile(sCleanPath)
+    
 
 
     print("compress file ?")
@@ -213,7 +247,7 @@ if __name__ == "__main__":
     if ( sYesNo == 1):
         #sFileServer = r"\\fileserver\Dep_AP\Project\SSD\temp\BinGrade"
         #sProgramimgPath = r"D:\\3S_PC\sourceCode\SSD\MP_UI\source_code\GIT_MP_UI_BIN_GRADE\v1.0\v1.0.2016.918_Temp1"
-        Achive_Folder_To_ZIP(sProgramimgPath, sFileServer, sVersion)
+        Achive_Folder_To_ZIP(sProgramimgPath, sServerSourceCode, sVersion)
     elif(sYesNo == 0):
         print("skip")
     else:        
@@ -224,28 +258,29 @@ if __name__ == "__main__":
     sYesNo = rawInputTest()
     if ( sYesNo == 1):
         sSrcPath = os.path.join(sProgramimgPath,"src\\bin")
-        copyTo3800(sSrcPath, sRemotePath, sVersion)
+        copyTo3800(sSrcPath, sServerBinary, sVersion)
     elif(sYesNo == 0):
         print("skip")
     else:        
         sys.exit(0)
 
 
-    print("copyReleaseNote ?")
+    print("copy change List ?")
     sYesNo = rawInputTest()
     if ( sYesNo == 1):
         s3800RemotePath = r"\\fileserver\3800\SW\SSD_MP_UI_EV\SSD_MP_UI_Release_Note.xls"
         sDep_APRemotePath = r"\\fileserver\Dep_AP\Project\SSD\MP_UI\source_code\SSD_MP_UI_Release_Note.xls"
-        sSrcPath = r"D:\\3S_PC\sourceCode\SSD\MP_UI\source_code\GIT_MP_UI\v1.0\SSD_MP_UI_Release_Note.xls"
-        copyReleaseNote(sSrcPath, s3800RemotePath)
-        copyReleaseNote(sSrcPath, sDep_APRemotePath)
+        sSrcPath = os.path.join(sProgramimgPath,"ChangeList.txt")
+        copyReleaseNote(sSrcPath, sServerSourceCode)
+        #copyReleaseNote(sSrcPath, sDep_APRemotePath)
     elif(sYesNo == 0):
         print("skip")
     else:        
         sys.exit(0)
 
-    
-
+    print("\n\n")
+    print("*** Excellent As You Are ***")
+    print("\n")
     #print("check D:\3S_PC\sourceCode\SSD\MP_UI\source_code\GIT_MP_UI\v1.0\v1.0.2016.918_Temp1\bin\SSDMP.exe date!!!")
 
     sys.exit(0)
